Update all csv files
This updates all csv files with a fresh export from the DLCS database. Many had been exported with an older version of the DLCSexport app; so column names and format might have changed.

Any manual changes made to the old csv files will be lost. Hopefully they are no longer necessary, but a history of those changes is still preserved in the git repository.